pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-spine-spine"
version = "0.1.0"
description = "PROJECT: SPINE — System spine do analizy RTG kręgosłupa"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }

dependencies = [
    # Deep Learning
    "torch>=2.2.0",
    "torchvision>=0.17.0",

    # Computer Vision
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.2.0",

    # Data & ML klasyczne
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",

    # Frontend (minimalny)
    "streamlit>=1.32.0",

    # Narzędzia
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]
# Opcjonalnie dla pełnego HRNet (wymaga torch już zainstalowanego)
hrnet = [
    "mmpose>=1.3.0",
    "mmengine>=0.10.0",
    "mmcv>=2.1.0",
]

[project.scripts]
spine-app = "spine_app:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
